[esbmc] split k-induction driver into parseoptions/
Moves doit_k_induction_parallel, is_base_case_violated,
does_forward_condition_hold, is_inductive_step_violated and
diagnose_unknown_properties, together with the file-local PROCESS_TYPE
enum and resultt struct they communicate over, into
src/esbmc/parseoptions/k_induction.cpp.

Verbatim move; the parallel path keeps the POSIX fork/pipe/waitpid
preamble. k-induction, k-induction-parallel and termination suites pass.
