+++ src/esbmc/CMakeLists.txt
@@ -46,6 +46,7 @@ add_executable(esbmc main.cpp esbmc_parseoptions.cpp
     parseoptions/bmc_strategy.cpp
     parseoptions/function_contracts.cpp
     parseoptions/goto_program.cpp
+    parseoptions/k_induction.cpp
     parseoptions/process_goto_program.cpp
     parseoptions/property_monitors.cpp
     bmc.cpp ranking_synthesis.cpp non_termination.cpp globals.cpp show_vcc.cpp options.cpp ${CMAKE_CURRENT_BINARY_DIR}/buildidobj.c)
